refactor: change healthbar folder path
