+++ public/examples/shoot/Monster/HealthBar.js
